+++ R/zzz.R
@@ -10,6 +10,7 @@
 #' @import mlr3
 #' @importFrom tools R_user_dir
 #' @importFrom withr with_seed
+#' @importFrom coro exhausted
 #'
 #' @section Options:
 #' * `mlr3torch.cache`: